Add arm64 build support within CI workflows
We are aware of the fact that rust already supports cross compiling.
However, for process parity and to avoid a multi-arch library
references mess, we decided to stick with docker buildx approach for
the moment.

This be can be changed in a future iteration, but for the moment it
solves the problem at hand.

Woth mentioning that now most of the build logic has been moved into
independent shell scripts, in hopes the developers make use of it.

Signed-off-by: Jonnatan Jossemar Cordero <jossemar.cordero@cto.ai>