Publish packaged artifacts
Since SDK bash binaries can be used on any Linux environment (that has a
compatible glib version), it is handy to be able to fetch the latest CI
artifacts for development purposes.

The artifacts published by the workflows have 24hrs lifespan and can be
only accessed by members of the GH organization.

Signed-off-by: Jonnatan Jossemar Cordero <jossemar.cordero@cto.ai>