PROD-3069: Build self-contained binaries w/musl
Up to this iteration we used to link our binaries to the GNU C library
(AKA glibc), which works fine for most of GNU distributions. However,
it is well known that managing said system dependency comes at a great
maitenance cost, so it would be best to ship self-contained binaries by
migrating into the musl C library.

Turns out, the only dependency that tied us to glibc was OpenSSL which
is no longer required for our HTTP client. So by changing some cargo
configurations we were able to do the switch we were aiming for.

This change came with the additional benefit of enabling cross
compiliation without relying on chip emulation, which is way faster! To
the point of reducing the CI time and billing by at least 5x.

Signed-off-by: Jonnatan Jossemar Cordero <jossemar.cordero@cto.ai>